Deprecate Honeybadger.wrap()
Honeybadger.wrap() exists in honeybadger-js (our client-side library)
for the purpose of wrapping callbacks in try/catch. In node This doesn't
take into account errors in asynchronous callbacks, and the synchronous
expectations of the current function (it must always re-throw) doesn't
play well with domains.

In the future I'd like to re-think how Honeybadger.wrap() should
*actually* work (if there is still a use case) and then either rewrite
or remove it.